do not render timeline cursor if no timelines are selected
